+++ Chip8Tests/Common/OPCodeTests.cs
@@ -7,7 +7,7 @@ public class OPCodeTests
     [Test]
     public void Constructor_Success()
     {
-        int testOPCode = 0x123F;
+        ushort testOPCode = 0x123F;
         OpCode opCode = new OpCode(testOPCode);
 
         Assert.That(opCode.X, Is.EqualTo(0x2), "X extraction did not work");
